requirements.txt
flask
pywebio
pillow
torch
torchvision
git+https://github.com/openai/CLIP.git
numpy
tqdm
